requirements.txt
PyPDF2>=3.0.1
pdf2image>=1.16.3
img2pdf>=0.4.4
fpdf>=1.7.2
PyMuPDF>=1.24.0
psutil>=5.9.5
pyttsx3>=2.90
Pillow>=10.0.0
pdfrw>=0.4
tqdm>=4.65.0
duckduckgo-search>=3.0.0
